only run dbus-launch if running on proxmox

+++ ohmydebn.sh
@@ -144,9 +144,11 @@ if [ $(dpkg -l | grep "^ii  mint-" | wc -l) -eq 0 ]; then
   sudo apt -y purge linuxmint-keyring
 fi
 
-display "tte rain" "Installing dbus-x11"
-sudo apt -y install dbus-x11
-export $(dbus-launch)
+if [ -f /usr/bin/pveversion ]; then
+  display "tte rain" "Installing dbus-x11"
+  sudo apt -y install dbus-x11
+  export $(dbus-launch)
+fi
 
 display "tte rain" "Changing desktop wallpaper"
 mkdir -p ~/.config/$PROJECT_LOWER/current/
